# release-2.1: opt: Increase performance of small queries