--- a/pkg/sql/opt/memo/logical_props_builder.go
+++ b/pkg/sql/opt/memo/logical_props_builder.go
@@ -33,7 +33,6 @@ var fdAnnID = opt.NewTableAnnID()
 type logicalPropsBuilder struct {
 	evalCtx         *tree.EvalContext
 	sb              statisticsBuilder
-	kb              keyBuffer
 	relationalAlloc []props.Relational
 	scalarAlloc     []props.Scalar
 }
@@ -181,7 +180,7 @@ func (b *logicalPropsBuilder) buildScanProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, md, &b.kb)
+	b.sb.init(b.evalCtx, md)
 	b.sb.buildScan(ev, relational)
 
 	return logical
@@ -217,7 +216,7 @@ func (b *logicalPropsBuilder) buildVirtualScanProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildVirtualScan(ev, relational)
 
 	return logical
@@ -279,7 +278,7 @@ func (b *logicalPropsBuilder) buildSelectProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildSelect(ev, relational)
 
 	return logical
@@ -369,7 +368,7 @@ func (b *logicalPropsBuilder) buildProjectProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildProject(ev, relational)
 
 	return logical
@@ -530,7 +529,7 @@ func (b *logicalPropsBuilder) buildJoinProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildJoin(ev, relational)
 
 	return logical
@@ -575,7 +574,7 @@ func (b *logicalPropsBuilder) buildIndexJoinProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, md, &b.kb)
+	b.sb.init(b.evalCtx, md)
 	b.sb.buildIndexJoin(ev, relational)
 
 	return logical
@@ -638,7 +637,7 @@ func (b *logicalPropsBuilder) buildGroupByProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildGroupBy(ev, relational)
 
 	return logical
@@ -696,7 +695,7 @@ func (b *logicalPropsBuilder) buildSetProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildSetOp(ev, relational)
 
 	return logical
@@ -737,7 +736,7 @@ func (b *logicalPropsBuilder) buildValuesProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildValues(ev, relational)
 
 	return logical
@@ -868,7 +867,7 @@ func (b *logicalPropsBuilder) buildLimitProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildLimit(ev, relational)
 
 	return logical
@@ -923,7 +922,7 @@ func (b *logicalPropsBuilder) buildOffsetProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildOffset(ev, relational)
 
 	return logical
@@ -962,7 +961,7 @@ func (b *logicalPropsBuilder) buildMax1RowProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildMax1Row(ev, relational)
 
 	return logical
@@ -1011,7 +1010,7 @@ func (b *logicalPropsBuilder) buildRowNumberProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildRowNumber(ev, relational)
 
 	return logical
@@ -1048,7 +1047,7 @@ func (b *logicalPropsBuilder) buildZipProps(ev ExprView) props.Logical {
 
 	// Statistics
 	// ----------
-	b.sb.init(b.evalCtx, ev.Metadata(), &b.kb)
+	b.sb.init(b.evalCtx, ev.Metadata())
 	b.sb.buildZip(ev, relational)
 
 	return logical