--- a/pkg/sql/opt/memo/private_storage.go
+++ b/pkg/sql/opt/memo/private_storage.go
@@ -606,16 +606,6 @@ func (kb *keyBuffer) writeColList(colList opt.ColList) {
 	}
 }
 
-// writeGroupList writes a series of varints, one for each column in the list,
-// in list order.
-func (kb *keyBuffer) writeGroupList(groupList []GroupID) {
-	var buf [10]byte
-	for _, col := range groupList {
-		cnt := binary.PutUvarint(buf[:], uint64(col))
-		kb.Write(buf[:cnt])
-	}
-}
-
 // writePhysProps writes the presentation columns, followed by the ordering
 // spec.
 func (kb *keyBuffer) writePhysProps(physical *props.Physical) {