… pkg/sql/opt/props/col_stats_map.go
… pkg/sql/opt/props/col_stats_map_test.go
… pkg/sql/opt/props/statistics.go